feat: add authentification requirement to all expenses routes

+++ src/controllers/expenses.controller.js
@@ -2,9 +2,10 @@ import { Op } from "sequelize";
 import db from "../models/index.js";
 
 const getAllExpenses = async (req, res) => {
+    const userUUID = req.user.id
     try {
         let { start, end, category, type } = req.query
-        let conditions = {}
+        let conditions = { user_id: userUUID }
         let errorList = []
 
         if (new Date(start) == 'Invalid Date') {
@@ -37,8 +38,8 @@ const getAllExpenses = async (req, res) => {
             }, {
                 association: 'user_fk',
                 attributes: ["email"]
-            }
-            ]
+            }],
+            order: [['expense_date', 'DESC']]
         })
         return res
             .status(200)
@@ -49,6 +50,7 @@ const getAllExpenses = async (req, res) => {
 }
 
 const getSpecificExpense = async (req, res) => {
+    const userUUID = req.user.id
     const { id } = req.params
 
     if (id.replaceAll('-', '').length !== 32) return res.status(400).json('Invalid expense ID')
@@ -56,7 +58,7 @@ const getSpecificExpense = async (req, res) => {
     try {
 
         const queryAnswer = await db.Expense.findOne({
-            where: { id: id },
+            where: { [Op.and]: { id: id, user_id: userUUID } },
             include: [{
                 association: 'category_fk',
                 attributes: ["name"]
@@ -77,27 +79,38 @@ const getSpecificExpense = async (req, res) => {
 
 // IMPORTANT: Need to get the user UUID from Authentification (requireAuth cannot pass it)
 const createExpense = async (req, res) => {
-    const { amount, expense_date, category_id, description, type, start_date, end_date } = req.body
-    if (!(amount && expense_date && category_id && description && type && start_date && end_date)) {
-        return res.status(400).json({ message: 'Missing field' })
+    let { amount, expense_date, category_id, description, is_recurrent, start_date, end_date } = req.body
+
+    if (!(amount && expense_date && category_id)) {
+        return res.status(400).json({ message: 'Missing field', field: Object.keys(req.body).filter(field => !req.body[field]) })
+    }
+
+    if (!is_recurrent) {
+        start_date = null
+        end_date = null
+    }
+
+    if (is_recurrent && !start_date) {
+        return res.status(400).json({ message: 'Recurrent expense must have a valid start date' })
     }
 
     try {
-        const userUUID = ''
+        const userUUID = req.user.id
 
-        const newExpense = await db.Expense.create({ user_id: userUUID, amount, expense_date, category_id, description, type, start_date, end_date })
+        const newExpense = await db.Expense.create({ user_id: userUUID, amount, expense_date, category_id, description, is_recurrent, start_date, end_date })
 
-        return res.status(201).json({ Expense: { user_id: userUUID, amount, expense_date, category_id, description, type, start_date, end_date }, message: 'Expense created' })
+        return res.status(201).json({ message: 'Expense created', Expense: { user_id: userUUID, amount, expense_date, category_id, description, is_recurrent, start_date, end_date } })
 
     } catch (err) { return res.status(500).json({ message: 'Server error', error: err.message }) }
 }
 
 const updateExpense = async (req, res) => {
+    const userUUID = req.user.id
     const { id } = req.params
 
     if (id.replaceAll('-', '').length !== 32) return res.status(400).json('Invalid expense ID')
 
-    const currentExpense = await db.Expense.findOne({ where: { id: id } });
+    const currentExpense = await db.Expense.findOne({ where: { [Op.and]: { id: id, user_id: userUUID } } });
     if (!currentExpense) return res.status(400).json('No match found')
 
     for (const field in currentExpense.toJSON()) {
@@ -112,12 +125,13 @@ const updateExpense = async (req, res) => {
 }
 
 const deleteExpense = async (req, res) => {
+    const userUUID = req.user.id
     const { id } = req.params;
 
     if (id.replaceAll('-', '').length !== 32) return res.status(400).json('Invalid expense ID')
 
     try {
-        const deleteExpense = await db.Expense.destroy({ where: { id: id } })
+        const deleteExpense = await db.Expense.destroy({ where: { [Op.and]: { id: id, user_id: userUUID } } })
         return (deleteExpense) ? res.status(200).json('Deletion successful') : res.status(400).json('No expense with such ID')
     } catch (err) {
         return res.status(500).json({ message: 'Failed to delete the record', error: err.message })
